pyproject.toml
[project]
name = "etl-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentic-doc>=0.3.1",
    "alembic>=1.16.4",
    "bs4>=0.0.2",
    "dask>=2025.7.0",
    "docling>=2.46.0",
    "ipykernel>=6.30.1",
    "langchain-text-splitters>=0.3.9",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "psycopg2>=2.9.10",
    "pyarrow>=21.0.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "weaviate-client>=4.16.7",
]
